split up parameters and remove ModelSetup
