+++ experiments/long_runs/soil.jl
@@ -84,33 +84,15 @@ function setup_model(FT, start_date, stop_date, domain, earth_param_set)
         max_wind_speed = 25.0,
         time_interpolation_method,
     )
-    spatially_varying_soil_params =
-        ClimaLand.ModelSetup.default_spatially_varying_soil_parameters(
-            subsurface_space,
-            surface_space,
-            FT,
-        )
-    (;
-        ν,
-        ν_ss_om,
-        ν_ss_quartz,
-        ν_ss_gravel,
-        hydrology_cm,
-        K_sat,
-        S_s,
-        θ_r,
-        PAR_albedo_dry,
-        NIR_albedo_dry,
-        PAR_albedo_wet,
-        NIR_albedo_wet,
-        f_max,
-    ) = spatially_varying_soil_params
+
+    (; ν_ss_om, ν_ss_quartz, ν_ss_gravel) =
+        ClimaLand.Soil.soil_composition_parameters(subsurface_space, FT)
+    (; ν, hydrology_cm, K_sat, θ_r) =
+        ClimaLand.Soil.soil_vangenuchten_parameters(subsurface_space, FT)
     soil_albedo = Soil.CLMTwoBandSoilAlbedo{FT}(;
-        PAR_albedo_dry,
-        NIR_albedo_dry,
-        PAR_albedo_wet,
-        NIR_albedo_wet,
+        ClimaLand.Soil.clm_soil_albedo_parameters(surface_space)...,
     )
+    S_s = ClimaCore.Fields.zeros(subsurface_space) .+ FT(1e-3)
     soil_params = Soil.EnergyHydrologyParameters(
         FT;
         ν,
@@ -123,12 +105,11 @@ function setup_model(FT, start_date, stop_date, domain, earth_param_set)
         θ_r,
         albedo = soil_albedo,
     )
-
     f_over = FT(3.28) # 1/m
     R_sb = FT(1.484e-4 / 1000) # m/s
     runoff_model = ClimaLand.Soil.Runoff.TOPMODELRunoff{FT}(;
         f_over = f_over,
-        f_max = f_max,
+        f_max = ClimaLand.Soil.topmodel_fmax(surface_space, FT),
         R_sb = R_sb,
     )
 
@@ -166,7 +147,7 @@ function setup_simulation(; greet = false)
         @info "Start Date: $start_date"
         @info "Stop Date: $stop_date"
     end
-    domain = ClimaLand.ModelSetup.global_domain(FT; context, nelements)
+    domain = ClimaLand.Domains.global_domain(FT; context, nelements)
     params = LP.LandParameters(FT)
     model = setup_model(FT, start_date, stop_date, domain, params)
     diagnostics = ClimaLand.default_diagnostics(
