+++ src/ClimaLand.jl
@@ -4,7 +4,7 @@ using DocStringExtensions
 using ClimaCore
 using LazyBroadcast: lazy
 import ClimaCore: Fields, Spaces
-
+include("Artifacts.jl")
 include("shared_utilities/Parameters.jl")
 import .Parameters as LP
 
@@ -20,7 +20,6 @@ import ClimaUtilities.SpaceVaryingInputs: SpaceVaryingInput
 
 import NCDatasets # Needed to load the ClimaUtilities.*VaryingInput
 using .Domains
-include("Artifacts.jl")
 include("shared_utilities/checkpoints.jl")
 include("shared_utilities/utils.jl")
 include("shared_utilities/models.jl")
@@ -425,43 +424,5 @@ include(joinpath("diagnostics", "Diagnostics.jl"))
 import .Diagnostics: default_diagnostics
 
 # Simulations
-include(joinpath("simulations", "ModelSetup.jl"))
 include(joinpath("simulations", "Simulations.jl"))
-
-"""
-    ClimaLand.global_domain
-
-a helper function which calls ModelSetup.global_domain from ClimaLand;
-implemented to prevent breaking changes in ClimaCoupler, but this can be removed
-at a later date.
-"""
-global_domain(args) = ModelSetup.global_domain(args...)
-
-"""
-    ClimaLand.default_spatially_varying_soil_parameters
-
-a helper function which calls ModelSetup.default_spatially_varying_soil_parameters from ClimaLand;
-implemented to prevent breaking changes in ClimaCoupler, but this can be removed
-at a later date.
-"""
-default_spatially_varying_soil_parameters(args...) =
-    ModelSetup.default_spatially_varying_soil_parameters(args...)
-
-"""
-    ClimaLand.clm_canopy_parameters
-
-a helper function which calls ModelSetup.clm_canopy_parameters from ClimaLand;
-implemented to prevent breaking changes in ClimaCoupler, but this can be removed
-at a later date.
-"""
-clm_canopy_parameters(args...) = ModelSetup.clm_canopy_parameters(args...)
-
-"""
-    ClimaLand.use_lowres_clm
-
-a helper function which calls ModelSetup.use_lowres_clm from ClimaLand;
-implemented to prevent breaking changes in ClimaCoupler, but this can be removed
-at a later date.
-"""
-use_lowres_clm(args...) = ModelSetup.use_lowres_clm(args...)
 end
