+++ src/shared_utilities/utils.jl
@@ -666,71 +666,6 @@ function NaNCheckCallback(
     SciMLBase.DiscreteCallback(cond, affect!)
 end
 
-apply_threshold(field, value) =
-    field > value ? eltype(field)(1) : eltype(field)(0)
-
-"""
-    landsea_mask(
-        surface_space;
-        filepath = ClimaLand.Artifacts.landseamask_file_path(;
-                 context = ClimaComms.context(surface_space),
-                 ),
-        threshold = 0.5,
-        regridder_type = :InterpolationsRegridder,
-        extrapolation_bc = (
-            Interpolations.Periodic(),
-            Interpolations.Flat(),
-            Interpolations.Flat(),
-        ),
-    )
-
-Reads in the default Clima land/sea mask, regrids to the `surface_space`, and
-treats any point with a land fraction < threshold as ocean.
-"""
-function landsea_mask(
-    surface_space;
-    filepath = ClimaLand.Artifacts.landseamask_file_path(;
-        context = ClimaComms.context(surface_space),
-    ),
-    threshold = 0.5,
-    regridder_type = :InterpolationsRegridder,
-    extrapolation_bc = (
-        Interpolations.Periodic(),
-        Interpolations.Flat(),
-        Interpolations.Flat(),
-    ),
-)
-    mask = SpaceVaryingInput(
-        filepath,
-        "landsea",
-        surface_space;
-        regridder_type,
-        regridder_kwargs = (; extrapolation_bc,),
-    )
-    binary_mask = apply_threshold.(mask, threshold)
-    return binary_mask
-end
-
-function landsea_mask(domain::Domains.AbstractDomain; kwargs...)
-    # average_horizontal_resolution_degrees returns a tuple with the resolution
-    # along the two directions, so we take the minimum
-    resolution_degrees = minimum(average_horizontal_resolution_degrees(domain))
-    if resolution_degrees > 1
-        resolution = "1deg"
-    else
-        resolution_arcsec = 3600resolution_degrees
-        # Pick the landsea mask at 60 arcseconds if the nodal distance is more than
-        # 120", otherwise pick the higher resolution
-        resolution = resolution_arcsec > 120 ? "60arcs" : "30arcs"
-    end
-
-    filepath = ClimaLand.Artifacts.landseamask_file_path(;
-        resolution,
-        context = ClimaComms.context(domain.space.surface),
-    )
-    return landsea_mask(domain.space.surface; filepath, kwargs...)
-end
-
 """
     ReportCallback(every_n_steps)
