+++ test/integrated/full_land_utils.jl
@@ -15,7 +15,7 @@ import ClimaLand
                        scalar_snow_params,
                        earth_param_set;
                        context = nothing,
-                       domain = ClimaLand.ModelSetup.global_domain(FT; context = context),
+                       domain = ClimaLand.Domains.global_domain(FT; context = context),
                        forcing = ClimaLand.prescribed_forcing_era5(ClimaLand.Artifacts.era5_land_forcing_data2008_path(; context),
                                                                    domain.space.surface,
                                                                    DateTime(2008),
@@ -41,7 +41,7 @@ function global_land_model(
     scalar_snow_params,
     earth_param_set;
     context = nothing,
-    domain = ClimaLand.ModelSetup.global_domain(FT; context = context),
+    domain = ClimaLand.Domains.global_domain(FT; context = context),
     forcing = ClimaLand.prescribed_forcing_era5(
         ClimaLand.Artifacts.era5_land_forcing_data2008_path(; context),
         domain.space.surface,
@@ -79,33 +79,14 @@ function global_land_model(
     surface_space = domain.space.surface
     subsurface_space = domain.space.subsurface
 
-    spatially_varying_soil_params =
-        ClimaLand.ModelSetup.default_spatially_varying_soil_parameters(
-            subsurface_space,
-            surface_space,
-            FT,
-        )
-    (;
-        ν,
-        ν_ss_om,
-        ν_ss_quartz,
-        ν_ss_gravel,
-        hydrology_cm,
-        K_sat,
-        S_s,
-        θ_r,
-        PAR_albedo_dry,
-        NIR_albedo_dry,
-        PAR_albedo_wet,
-        NIR_albedo_wet,
-        f_max,
-    ) = spatially_varying_soil_params
-    albedo = Soil.CLMTwoBandSoilAlbedo{FT}(;
-        PAR_albedo_dry,
-        NIR_albedo_dry,
-        PAR_albedo_wet,
-        NIR_albedo_wet,
+    (; ν_ss_om, ν_ss_quartz, ν_ss_gravel) =
+        ClimaLand.Soil.soil_composition_parameters(subsurface_space, FT)
+    (; ν, hydrology_cm, K_sat, θ_r) =
+        ClimaLand.Soil.soil_vangenuchten_parameters(subsurface_space, FT)
+    soil_albedo = Soil.CLMTwoBandSoilAlbedo{FT}(;
+        ClimaLand.Soil.clm_soil_albedo_parameters(surface_space)...,
     )
+    S_s = ClimaCore.Fields.zeros(subsurface_space) .+ FT(1e-3)
     soil_params = Soil.EnergyHydrologyParameters(
         FT;
         ν,
@@ -116,28 +97,23 @@ function global_land_model(
         K_sat,
         S_s,
         θ_r,
-        albedo,
+        albedo = soil_albedo,
     )
+    f_over = FT(3.28) # 1/m
+    R_sb = FT(1.484e-4 / 1000) # m/s
     runoff_model = ClimaLand.Soil.Runoff.TOPMODELRunoff{FT}(;
         f_over = f_over,
-        f_max = f_max,
+        f_max = ClimaLand.Soil.topmodel_fmax(surface_space, FT),
         R_sb = R_sb,
     )
 
     # Spatially varying canopy parameters from CLM
-    clm_parameters = ClimaLand.ModelSetup.clm_canopy_parameters(surface_space)
-    (;
-        Ω,
-        rooting_depth,
-        is_c3,
-        Vcmax25,
-        g1,
-        G_Function,
-        α_PAR_leaf,
-        τ_PAR_leaf,
-        α_NIR_leaf,
-        τ_NIR_leaf,
-    ) = clm_parameters
+    g1 = ClimaLand.Canopy.clm_medlyn_g1(surface_space)
+    rooting_depth = ClimaLand.Canopy.clm_rooting_depth(surface_space)
+    (; is_c3, Vcmax25) =
+        ClimaLand.Canopy.clm_photosynthesis_parameters(surface_space)
+    (; Ω, G_Function, α_PAR_leaf, τ_PAR_leaf, α_NIR_leaf, τ_NIR_leaf) =
+        ClimaLand.Canopy.clm_canopy_radiation_parameters(surface_space)
 
     conductivity_model =
         Canopy.PlantHydraulics.Weibull{FT}(K_sat_plant, ψ63, Weibull_param)
