+++ test/runtests.jl
@@ -84,6 +84,10 @@ end
     include("standalone/Soil/conservation.jl")
 end
 
+@safetestset "Soil spatial parameters" begin
+    include("standalone/Soil/spatial_parameters.jl")
+end
+
 # Standalone Surface Water model tests
 @safetestset "Pond module tests" begin
     include("standalone/SurfaceWater/pond_test.jl")
@@ -106,6 +110,10 @@ end
     include("standalone/Vegetation/conservation.jl")
 end
 
+@safetestset "Canopy spatial parameters" begin
+    include("standalone/Vegetation/spatial_parameters.jl")
+end
+
 # Integrated LSM tests
 @safetestset "Integrated LSM unit tests" begin
     include("integrated/lsms.jl")
@@ -124,11 +132,6 @@ end
 end
 @safetestset "Full land" begin
     include("integrated/full_land.jl")
-end
-
-# Simulations
-@safetestset "Default parameters" begin
-    include("simulations/spatial_parameters.jl")
 end
 
 # Diagnostics
